Migrate GetAppLanguagesInterfaceStringsRepositoryTests to Swift Tests
